Update code for responsive designs
Added support for responsive layouts. Normally in those cases when the screen width changes so does the element's width and therefore height (because of the content in it).

GSAP instances don't recognize this changes so they have to be created again. Because of that the code gets the current progress of the instance, the play state and the direction, just in case a resize is triggered while the animation is playing/reversing